Change mtime/atime of index.html
To a date guaranteed to be before the mtime of any files generated
in the main function. This allows identifying all files which have not
been touched by the main function, i.e. removed from the repository by
comparing the mtime of these files with the mtime of index.html. For
example, using:

	find . -not -newer index.html