Add small Javascript function to highlight a range of lines
With vanilla CSS only a single line can be highlighted. This commit adds
a small Javascript function to highlight a range of lines with the same
syntax used by GitLab/GitHub. The change is explicitly designed in a way
that highlighting a single line without Javascript still works fine and
Javascript remains entirely optional.