css: Fix scrollbar in overflowing <pre>
This was introduced in 9538a5a630996a7945cdd1577e739d258a0b47c5. Since
this commit, the <pre> would not include a scrollbar even if the <pre>
was overflowing.

With this change, if <pre> is not overflowing the entirety of the line
is highlighted including any remaining space on the right. When the
<pre> is overflowing and a scrollbar is added, only the text in the line
is highlighted. I am not sure if it is possible to also highlight the
entire width when the scrollbar is present but this is decent enough for
now I think.