depp-index: Support stripping .git suffix when linking to repository
This is useful when serving both the repository itself and the generated
HTML files from the same directory over HTTP. In such a scenario, the
repository may be cloned over https://git.example.org/foo.git the HTML
would then be served via https://git.example.org/foo.

In order to support these sorts of setups, this commit adds an
additional flag for optionally stripping the .git suffix on the index
page when linking to the Git repository HTML. By default, the .git
suffix is retained thus this change is backwards compatible.