gitweb: In incremental builds, create empty dirs for renames
Without this patch, when renaming a file, a new empty directory may
be introduced. For example, when moving files from app/* to src/Foo/*.
In this example, src/Foo would be an empty directory and, hence, would
not show up in .FilePatches(); therefore, we need to iterate over each
element of the new path and check for empty directories explicitly.